apps/worker/pipeline/mcp-server/requirements.txt
mcp>=1.0.0,<2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
google-genai>=1.0.0
